ESP32 as WiFi portal for Teensy 4.0.
This project, compiled under Arduino, is a data acquisition system serving the data via webpage, and allowing the webpage to control and LED on the prototype.   It uses separate dedicated microprocessors as a front-end, and back-end for a reason.  The front-end processor in this implementation is a Teensy 4.0, which gives it a lot of I/O capability, far more than the ESP32 alone.    The Teensy also has the speed necessary to do a significant amount of data processing maths.  The "back-end" here is an ESP32, in this case, a Heltec Lora/Wifi/BT module with an OLED display.    The front and back end are connected by a simple logic level serial link to allow any sort of protocol to control transfer of the processed front end data to the ESP32 back-end which provides a simple webpage interface.   In the initial commit, the webpage allows display of temperature, humidity and barometric pressure collected from a BME380 sensor interfaced to the Teensy 4.0 front end.   
Note both the front and back end here are actually parts of a front end, i.e. the device interfaced to the physical world.    Also note the prototype includes control of an RC servo motor as a fun way to display the temperature.